+++ server/src/main/java/org/allaymc/server/network/protocol/PacketEncoder.java
@@ -134,6 +134,11 @@ public abstract class PacketEncoder {
         return null;
     }
 
+    /** Encodes the jigsaw structure rules */
+    public JigsawStructureDataPacket encodeJigsawStructureData() {
+        return null;
+    }
+
     /** Encodes metadata for the resource packs offered by the server. */
     public ResourcePacksInfoPacket encodeResourcePacksInfo() {
         return null;