+++ server/src/test/java/org/allaymc/server/network/protocol/ProtocolInheritanceTest.java
@@ -6,6 +6,10 @@ import org.allaymc.server.network.protocol.v2168.PacketEncoder_v2168;
 import org.allaymc.server.network.protocol.v2168.Protocol_v2168;
 import org.allaymc.server.network.protocol.v2169.PacketEncoder_v2169;
 import org.allaymc.server.network.protocol.v2169.Protocol_v2169;
+import org.allaymc.server.network.protocol.v2192.PacketEncoder_v2192;
+import org.allaymc.server.network.protocol.v2192.Protocol_v2192;
+import org.allaymc.server.network.protocol.v2193.PacketEncoder_v2193;
+import org.allaymc.server.network.protocol.v2193.Protocol_v2193;
 import org.allaymc.server.network.protocol.v766.PacketEncoder_v766;
 import org.allaymc.server.network.protocol.v766.PacketEncoder_v766_NetEase;
 import org.allaymc.server.network.protocol.v766.Protocol_v766;
@@ -66,7 +70,9 @@ class ProtocolInheritanceTest {
                 Protocol_v975.class,
                 Protocol_v1001.class,
                 Protocol_v2168.class,
-                Protocol_v2169.class
+                Protocol_v2169.class,
+                Protocol_v2192.class,
+                Protocol_v2193.class
         ));
     }
 
@@ -86,7 +92,9 @@ class ProtocolInheritanceTest {
                 PacketEncoder_v975.class,
                 PacketEncoder_v1001.class,
                 PacketEncoder_v2168.class,
-                PacketEncoder_v2169.class
+                PacketEncoder_v2169.class,
+                PacketEncoder_v2192.class,
+                PacketEncoder_v2193.class
         ));
     }
 
@@ -115,9 +123,11 @@ class ProtocolInheritanceTest {
                 new Protocol_v975(),
                 new Protocol_v1001(),
                 new Protocol_v2168(),
-                new Protocol_v2169()
+                new Protocol_v2169(),
+                new Protocol_v2192(),
+                new Protocol_v2193()
         );
-        var expectedVersions = List.of(818, 819, 827, 844, 859, 860, 898, 924, 944, 975, 1001, 2168, 2169);
+        var expectedVersions = List.of(818, 819, 827, 844, 859, 860, 898, 924, 944, 975, 1001, 2168, 2169, 2192, 2193);
 
         for (int index = 0; index < international.size(); index++) {
             var protocol = international.get(index);